requirements.txt
numpy
Pillow
torch
torchvision
tqdm
scikit-learn
PyYAML
